+++ shaders/gui_uv.vert
@@ -1,6 +1,6 @@
 #version 450
 
-layout(location=0) in uvec2 i_pos;
+layout(location=0) in vec2 i_pos;
 layout(location=1) in vec2 i_uv;
 
 layout(location=0) out vec2 o_uv;
